+++ src/extension.ts
@@ -28,6 +28,7 @@ export async function activate(context) {
         vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.add_missing_prop`, async () => await _refactor.addMethodMissingProperty()),
         // new
         vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.add_constructor`, async () => await _refactor.addConstructor()),
+        vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.add_invoke`, async () => await _refactor.addInvoke()),
         vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.add_new_property`, async () => await _refactor.addNewProperty()),
         // providers
         vscode.languages.registerCodeActionsProvider('php', new CodeAction()),